CASMPET-2399: Change webhook to not override fsGroup
After upgrading to istio 1.5.4 we found that postgres pods wouldn't
start because they couldn't read their tls files. This was because
the patroni operator starts these pods with a specific fsGroup and
now the istio sidecar webhook overwrites the pod fsGroup so
postgres can't read the secret.

The workaround is to modify the webhook to not overwrite the
fsGroup.

Reference this istio issue: https://github.com/istio/istio/issues/22025

This comment indicates that somebody has tried this and it has worked
for them: https://github.com/istio/istio/issues/22025#issuecomment-604196723