CASMPET-3059 - Remove duplicated metrics when cronjob-pod talks to istio-proxy sidecar
This re-implements 12a7e75160d from shasta 1.3 istio 1.5.4.
In that case the only change was to assets.gen.go but I think that
the community switched to generating that file rather than checking
it in, so I modified the source files and re-ran

BUILD_WITH_CONTAINER=1 make gen-charts

To create the assets.gen.go.