CASMPET-3736: Also publish the proxyv2 image
The VARIANT environment variable is now set so that the images
will have a tag of 1.6.13-cray1. The tag needs to change since we
can't re-use 1.6.13, and 1.6.14 is the next istio release so that
would be confusing.

In shasta 1.4 / csm 0.9 we started shipping the proxyv2 image
because we needed to patch all the images to include the fix
for CVE-2021-3156 [1].

[1] https://ubuntu.com/security/CVE-2021-3156

The CVE has been fixed in the upstream base image now so we no
longer need to build the proxyv2 image for this, but since we're
already building and using our own image we might as well continue
doing that in case we need to patch it in the future.