CASMPET-3767: Build both distroful and distroless
The Jenkinsfile is updated to build both distroful and distroless
images. The snyk tool is reporting several security vulnerabilities
that are in the distro so running distroless will remove those
vulnerabilities.

The 'Tag and Save image' stage is removed, DST said this was no longer
needed and it uses dtr which is going away.